requirements.txt
python-yelp-v2
foursquare